battery-status/battery-discharging-04-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Battery Status Test: BatteryManager - level (discharging)</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/battery-status/">
<meta name="flags" content="interact">
<meta name="assert" content="The level attribute must be set to the current battery level scaled from 0 to 1.0">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <h2>Preconditions</h2>
  <ol>
    <li>
      The device is plugged in to the charger before this test is run.
    </li>
    <li>
      The battery must neither be empty or full, nor reach empty or full capacity during the test.
    </li>
  </ol>
</body>
<script>

async_test(function (t) {
  navigator.getBattery().then(function (battery) {
      t.step(function () {
        assert_greater_than(battery.level, 0);
        assert_less_than(battery.level, 1.0);
      });
      t.done();
    }, function (error) {
      t.step(function () {
        assert_unreached(error.message);
      });
      t.done();
    }
  )
}, document.title);

</script>
